Split up Commandry into registration and dispatching
